+++ NEWS.md
@@ -1,6 +1,7 @@
 # crew 0.3.0.9000 (development)
 
 * Remove temporary controller summary workaround that ensured CRAN compatibility with `targets` 1.1.3.
+* Add a `map()` method to the controller class for functional programming (#93).
 
 # crew 0.3.0